+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 7.1.0 IN PROGRESS
 
 * Refactor connect to use a functional component. Refs STCON-132.
+* Move `initResources` to `useEffect` to avoid react warning. Fixes STCON-133.
 
 ## [7.0.0](https://github.com/folio-org/stripes-connect/tree/v7.0.0) (2021-09-26)
 [Full Changelog](https://github.com/folio-org/stripes-connect/compare/v6.2.0...v7.0.0)